django_test_work/requirements.txt
pymongo==3.12.3
django
djongo
pytz
schedule
requests
django-background-tasks